+++ tests/h2integrate/test_all_examples.py
@@ -322,6 +322,7 @@ def test_wind_h2_opt_example(subtests):
     assert len(cases) > 1, "Not enough cases recorded in SQL file."
 
     # Get initial and final LCOH values
+
     initial_lcoh = cases[0].outputs["finance_subgroup_hydrogen.LCOH"][0]
     final_lcoh = cases[-1].outputs["finance_subgroup_hydrogen.LCOH"][0]
 
@@ -466,7 +467,7 @@ def test_hybrid_energy_plant_example(subtests):
 
     # Subtests for checking specific values
     with subtests.test("Check LCOE"):
-        assert model.prob.get_val("finance_subgroup_default.LCOE", units="USD/MW/h")[0] < 83.2123
+        assert model.prob.get_val("finance_subgroup_default.LCOE", units="USD/(MW*h)")[0] < 83.2123
 
 
 def test_asu_example(subtests):
@@ -506,19 +507,28 @@ def test_hydrogen_dispatch_example(subtests):
     with subtests.test("Check LCOE"):
         assert (
             pytest.approx(
-                model.prob.get_val("finance_subgroup_elec.LCOE", units="USD/MW/h")[0],
+                model.prob.get_val("finance_subgroup_electricity.LCOE", units="USD/(MW*h)")[0],
                 rel=1e-5,
             )
-            == 105.75426
+            == 59.0962072084844
         )
 
-    with subtests.test("Check LCOH"):
+    with subtests.test("Check all h2 LCOH"):
         assert (
             pytest.approx(
-                model.prob.get_val("finance_subgroup_h2.LCOH", units="USD/kg")[0],
+                model.prob.get_val("finance_subgroup_all_hydrogen.LCOH", units="USD/kg")[0],
                 rel=1e-5,
             )
-            == 5.6540332
+            == 5.360810057454742
+        )
+
+    with subtests.test("Check dispatched h2 LCOH"):
+        assert (
+            pytest.approx(
+                model.prob.get_val("finance_subgroup_dispatched_hydrogen.LCOH", units="USD/kg")[0],
+                rel=1e-5,
+            )
+            == 8.371194137479359
         )
 
 
@@ -679,7 +689,7 @@ def test_wind_solar_electrolyzer_example(subtests):
     with subtests.test("Check LCOE"):
         assert (
             pytest.approx(
-                model.prob.get_val("finance_subgroup_electricity.LCOE", units="USD/MW/h")[0],
+                model.prob.get_val("finance_subgroup_electricity.LCOE", units="USD/(MW*h)")[0],
                 rel=1e-5,
             )
             == 53.9306558
@@ -716,7 +726,7 @@ def test_electrolyzer_om_example(subtests):
 
     model.run()
 
-    lcoe = model.prob.get_val("finance_subgroup_electricity.LCOE", units="USD/MW/h")[0]
+    lcoe = model.prob.get_val("finance_subgroup_electricity.LCOE", units="USD/(MW*h)")[0]
     lcoh_with_lcoh_finance = model.prob.get_val(
         "finance_subgroup_hydrogen.LCOH_lcoh_financials", units="USD/kg"
     )[0]
@@ -724,11 +734,11 @@ def test_electrolyzer_om_example(subtests):
         "finance_subgroup_hydrogen.LCOH_lcoe_financials", units="USD/kg"
     )[0]
     with subtests.test("Check LCOE"):
-        assert pytest.approx(lcoe, rel=1e-5) == 39.98869
+        assert pytest.approx(lcoe, rel=1e-4) == 39.98869
     with subtests.test("Check LCOH with lcoh_financials"):
-        assert pytest.approx(lcoh_with_lcoh_finance, rel=1e-5) == 13.0954678
+        assert pytest.approx(lcoh_with_lcoh_finance, rel=1e-4) == 13.0954678
     with subtests.test("Check LCOH with lcoe_financials"):
-        assert pytest.approx(lcoh_with_lcoe_finance, rel=1e-5) == 8.00321771
+        assert pytest.approx(lcoh_with_lcoe_finance, rel=1e-4) == 8.00321771
 
 
 def test_wombat_electrolyzer_example(subtests):
@@ -741,10 +751,10 @@ def test_wombat_electrolyzer_example(subtests):
     model.run()
 
     lcoe_with_profast_model = model.prob.get_val(
-        "finance_subgroup_electricity_profast.LCOE", units="USD/MW/h"
+        "finance_subgroup_electricity_profast.LCOE", units="USD/(MW*h)"
     )[0]
     lcoe_with_custom_model = model.prob.get_val(
-        "finance_subgroup_electricity_custom.LCOE", units="USD/MW/h"
+        "finance_subgroup_electricity_custom.LCOE", units="USD/(MW*h)"
     )[0]
 
     lcoh_with_custom_model = model.prob.get_val(
@@ -764,9 +774,94 @@ def test_wombat_electrolyzer_example(subtests):
         assert pytest.approx(lcoe_with_profast_model, rel=1e-5) == 59.0962084
 
 
-def test_wind_battery_dispatch_example(subtests):
+def test_pyomo_heuristic_dispatch_example(subtests):
     # Change the current working directory to the example's directory
-    os.chdir(EXAMPLE_DIR / "19_wind_battery_dispatch")
+    os.chdir(EXAMPLE_DIR / "18_pyomo_heuristic_dispatch")
+
+    # Create a H2Integrate model
+    model = H2IntegrateModel(Path.cwd() / "pyomo_heuristic_dispatch.yaml")
+
+    demand_profile = np.ones(8760) * 50.0
+
+    # TODO: Update with demand module once it is developed
+    model.setup()
+    model.prob.set_val("battery.electricity_demand", demand_profile, units="MW")
+
+    # Run the model
+    model.run()
+
+    model.post_process()
+
+    # Test battery storage functionality
+    # SOC should stay within configured bounds (10% to 90%)
+    # Due to pysam simulation, bounds may not be fully respected,
+    # but should not exceed the upper bound more than 4% SOC
+    # and the lower bound more than 1% SOC
+    soc = model.prob.get_val("battery.SOC")
+    with subtests.test("Check battery SOC lower bound"):
+        assert all(soc >= 9.0)
+    with subtests.test("Check battery SOC upper bound"):
+        assert all(soc <= 94.0)
+
+    with subtests.test("Check wind generation out of the wind plant"):
+        # Wind should generate some electricity
+        wind_electricity = model.prob.get_val("wind.electricity_out")
+        assert wind_electricity.sum() > 0
+        # Wind electricity should match battery input (direct connection)
+    with subtests.test("Check wind generation in to battery"):
+        battery_electricity_in = model.prob.get_val("battery.electricity_in")
+        assert wind_electricity.sum() == pytest.approx(battery_electricity_in.sum(), rel=1e-6)
+
+    with subtests.test("Check demand satisfaction"):
+        electricity_out = model.prob.get_val("battery.electricity_out", units="MW")
+        # Battery output should try to meet the 50 MW constant demand
+        # Average output should be close to demand when there's sufficient generation
+        assert electricity_out.mean() >= 45  # MW
+
+    # Subtest for LCOE
+    with subtests.test("Check all LCOE value"):
+        lcoe = model.prob.get_val("finance_subgroup_all_electricity.LCOE")[0]
+        assert lcoe == pytest.approx(0.08157197567200995, rel=1e-6)
+
+    with subtests.test("Check dispatched LCOE value"):
+        lcoe = model.prob.get_val("finance_subgroup_dispatched_electricity.LCOE")[0]
+        assert lcoe == pytest.approx(0.5975902853904799, rel=1e-6)
+
+    # Subtest for total electricity produced
+    with subtests.test("Check total electricity produced"):
+        total_electricity = model.prob.get_val(
+            name="finance_subgroup_all_electricity.electricity_sum.total_electricity_produced",
+            units="MW*h/year",
+        )[0]
+        assert total_electricity == pytest.approx(3125443.1089529935, rel=1e-6)
+
+    # Subtest for electricity unused_commodity
+    with subtests.test("Check electricity unused commodity"):
+        electricity_unused_commodity = np.linalg.norm(
+            model.prob.get_val("battery.unused_electricity_out", units="MW")
+        )
+        assert electricity_unused_commodity == pytest.approx(36590.067573337095, rel=1e-6)
+
+    # Subtest for unmet demand
+    with subtests.test("Check electricity unmet demand"):
+        electricity_unmet_demand = np.linalg.norm(
+            model.prob.get_val("battery.unmet_electricity_demand_out", units="MW")
+        )
+        assert electricity_unmet_demand == pytest.approx(711.1997294551337, rel=1e-6)
+
+    # check that error is raised when incorrect tech_name is given
+    with subtests.test("Check incorrect tech_name error"):
+        expected_error = (
+            r"tech_name in control_parameters \(wrong_tech_name\) must match "
+            r"the top-level name of the tech group \(battery\)"
+        )
+        with pytest.raises(ValueError, match=expected_error):
+            H2IntegrateModel(Path.cwd() / "pyomo_heuristic_dispatch_error_for_testing.yaml")
+
+
+def test_simple_dispatch_example(subtests):
+    # Change the current working directory to the example's directory
+    os.chdir(EXAMPLE_DIR / "19_simple_dispatch")
 
     # Create a H2Integrate model
     model = H2IntegrateModel(Path.cwd() / "wind_battery_dispatch.yaml")
@@ -799,8 +894,22 @@ def test_wind_battery_dispatch_example(subtests):
 
     # Subtest for LCOE
     with subtests.test("Check LCOE value"):
-        lcoe = model.prob.get_val("finance_subgroup_electricity.LCOE")[0]
+        lcoe = model.prob.get_val("finance_subgroup_electricity.LCOE_all_electricity_profast")[0]
         assert pytest.approx(lcoe, rel=1e-6) == 0.07801723344476236
+
+    # Subtest for NPV
+    with subtests.test("Check NPV value"):
+        npv = model.prob.get_val(
+            "finance_subgroup_electricity.NPV_electricity_all_electricity_npv"
+        )[0]
+        assert pytest.approx(npv, rel=1e-6) == 3791194.71
+
+    # Subtest for ProFAST NPV
+    with subtests.test("Check NPV value"):
+        npv = model.prob.get_val(
+            "finance_subgroup_electricity.NPV_electricity_all_electricity_profast_npv"
+        )[0]
+        assert pytest.approx(npv, rel=1e-6) == 7518969.18
 
     # Subtest for total electricity produced
     with subtests.test("Check total electricity produced"):
@@ -809,14 +918,192 @@ def test_wind_battery_dispatch_example(subtests):
         )[0]
         assert pytest.approx(total_electricity, rel=1e-6) == 62797265.9296355
 
-    # Subtest for electricity curtailed
-    with subtests.test("Check electricity curtailed"):
-        electricity_curtailed = np.linalg.norm(model.prob.get_val("battery.electricity_curtailed"))
-        assert pytest.approx(electricity_curtailed, rel=1e-6) == 412531.73840450746
-
-    # Subtest for missed load
-    with subtests.test("Check electricity missed load"):
-        electricity_missed_load = np.linalg.norm(
-            model.prob.get_val("battery.electricity_missed_load")
+    # Subtest for electricity unused_commodity
+    with subtests.test("Check electricity unused commodity"):
+        electricity_unused_commodity = np.linalg.norm(
+            model.prob.get_val("battery.electricity_unused_commodity")
         )
-        assert pytest.approx(electricity_missed_load, rel=1e-6) == 165604.70758669
+        assert pytest.approx(electricity_unused_commodity, rel=1e-6) == 412531.73840450746
+
+    # Subtest for unmet demand
+    with subtests.test("Check electricity unmet demand"):
+        electricity_unmet_demand = np.linalg.norm(
+            model.prob.get_val("battery.electricity_unmet_demand")
+        )
+        assert pytest.approx(electricity_unmet_demand, rel=1e-6) == 165604.70758669
+
+    # Subtest for total electricity produced from wind, should be equal to total
+    # electricity produced from finance_subgroup_electricity
+    with subtests.test("Check total electricity produced from wind"):
+        wind_electricity_finance = model.prob.get_val(
+            "finance_subgroup_wind.electricity_sum.total_electricity_produced", units="kW*h/year"
+        )[0]
+        assert pytest.approx(wind_electricity_finance, rel=1e-6) == total_electricity
+
+    with subtests.test("Check total electricity produced from wind compared to wind aep"):
+        wind_electricity_performance = np.sum(
+            model.prob.get_val("wind.electricity_out", units="kW")
+        )
+        assert pytest.approx(wind_electricity_performance, rel=1e-6) == wind_electricity_finance
+
+    # Subtest for total electricity produced from battery, should be equal
+    # to sum of "battery.electricity_out"
+    with subtests.test("Check total electricity produced from battery"):
+        battery_electricity_finance = model.prob.get_val(
+            "finance_subgroup_battery.electricity_sum.total_electricity_produced", units="MW*h/year"
+        )[0]
+        battery_electricity_performance = np.sum(
+            model.prob.get_val("battery.electricity_out", units="MW")
+        )
+        assert (
+            pytest.approx(battery_electricity_finance, rel=1e-6) == battery_electricity_performance
+        )
+
+    wind_lcoe = model.prob.get_val("finance_subgroup_wind.LCOE_wind_only", units="USD/(MW*h)")[0]
+    battery_lcoe = model.prob.get_val(
+        "finance_subgroup_battery.LCOE_battery_included", units="USD/(MW*h)"
+    )[0]
+    electricity_lcoe = model.prob.get_val(
+        "finance_subgroup_electricity.LCOE_all_electricity_profast", units="USD/(MW*h)"
+    )[0]
+
+    with subtests.test("Check electricity LCOE is greater than wind LCOE"):
+        assert electricity_lcoe > wind_lcoe
+
+    with subtests.test("Check battery LCOE is greater than electricity LCOE"):
+        assert battery_lcoe > electricity_lcoe
+
+    with subtests.test("Check battery LCOE"):
+        assert pytest.approx(battery_lcoe, rel=1e-6) == 131.781997
+
+    with subtests.test("Check wind LCOE"):
+        assert pytest.approx(wind_lcoe, rel=1e-6) == 58.8248
+
+    with subtests.test("Check electricity LCOE"):
+        assert pytest.approx(electricity_lcoe, rel=1e-6) == 78.01723
+
+
+def test_csvgen_design_of_experiments(subtests):
+    os.chdir(EXAMPLE_DIR / "20_solar_electrolyzer_doe")
+
+    # Create a H2Integrate model
+
+    with pytest.raises(UserWarning) as excinfo:
+        model = H2IntegrateModel(Path.cwd() / "20_solar_electrolyzer_doe.yaml")
+        assert "There may be issues with the csv file csv_doe_cases.csv" in str(excinfo.value)
+
+    # Run the model
+    from hopp.utilities.utilities import load_yaml
+
+    from h2integrate.core.utilities import check_file_format_for_csv_generator
+    from h2integrate.core.dict_utils import update_defaults
+    from h2integrate.core.inputs.validation import write_yaml, load_driver_yaml
+
+    # load the driver config file
+    driver_config = load_driver_yaml("driver_config.yaml")
+    # specify the filepath to the csv file
+    csv_fpath = Path(driver_config["driver"]["design_of_experiments"]["filename"]).absolute()
+    # run the csv checker method, we want it to write the csv file to a new filepath so
+    # set overwrite_file=False
+    new_csv_filename = check_file_format_for_csv_generator(
+        csv_fpath, driver_config, check_only=False, overwrite_file=False
+    )
+
+    # update the csv filename in the driver config dictionary
+    updated_driver = update_defaults(driver_config["driver"], "filename", new_csv_filename.name)
+    driver_config["driver"].update(updated_driver)
+
+    # save the updated driver to a new file
+    new_driver_fpath = Path.cwd() / "driver_config_test.yaml"
+    new_toplevel_fpath = Path.cwd() / "20_solar_electrolyzer_doe_test.yaml"
+    write_yaml(driver_config, new_driver_fpath)
+
+    # update the driver config filename in the top-level config
+    main_config = load_yaml("20_solar_electrolyzer_doe.yaml")
+    main_config["driver_config"] = new_driver_fpath.name
+
+    # save the updated top-level config file to a new file
+    write_yaml(main_config, new_toplevel_fpath)
+
+    model = H2IntegrateModel(new_toplevel_fpath)
+    model.run()
+
+    sql_fpath = Path.cwd() / "ex_20_out" / "cases.sql"
+    cr = om.CaseReader(str(sql_fpath))
+    cases = list(cr.get_cases())
+
+    with subtests.test("Check solar capacity in case 0"):
+        assert pytest.approx(cases[0].get_val("solar.capacity_kWdc", units="MW"), rel=1e-6) == 25.0
+    with subtests.test("Check solar capacity in case 9"):
+        assert (
+            pytest.approx(cases[-1].get_val("solar.capacity_kWdc", units="MW"), rel=1e-6) == 500.0
+        )
+
+    with subtests.test("Check electrolyzer capacity in case 0"):
+        assert (
+            pytest.approx(
+                cases[0].get_val("electrolyzer.electrolyzer_size_mw", units="MW"), rel=1e-6
+            )
+            == 10.0 * 5
+        )
+
+    with subtests.test("Check electrolyzer capacity in case 9"):
+        assert (
+            pytest.approx(
+                cases[-1].get_val("electrolyzer.electrolyzer_size_mw", units="MW"), rel=1e-6
+            )
+            == 10.0 * 10
+        )
+
+    min_lcoh_val = 100000.0
+    min_lcoh_case_num = 0
+    for i, case in enumerate(cases):
+        lcoh = case.get_val("finance_subgroup_hydrogen.LCOH_optimistic", units="USD/kg")[0]
+        if lcoh < min_lcoh_val:
+            min_lcoh_val = np.min([lcoh, min_lcoh_val])
+            min_lcoh_case_num = i
+
+    with subtests.test("Min LCOH value"):
+        assert pytest.approx(min_lcoh_val, rel=1e-6) == 4.468258
+
+    with subtests.test("Min LCOH case number"):
+        assert min_lcoh_case_num == 6
+
+    with subtests.test("Min LCOH case LCOH value"):
+        assert (
+            pytest.approx(
+                cases[min_lcoh_case_num].get_val(
+                    "finance_subgroup_hydrogen.LCOH_optimistic", units="USD/kg"
+                ),
+                rel=1e-6,
+            )
+            == min_lcoh_val
+        )
+
+    with subtests.test("Min LCOH case has lower LCOH than other cases"):
+        for i, case in enumerate(cases):
+            lcoh_case = case.get_val("finance_subgroup_hydrogen.LCOH_optimistic", units="USD/kg")
+            if i != min_lcoh_case_num:
+                assert lcoh_case > min_lcoh_val
+
+    with subtests.test("Min LCOH solar capacity"):
+        assert (
+            pytest.approx(
+                cases[min_lcoh_case_num].get_val("solar.capacity_kWdc", units="MW"), rel=1e-6
+            )
+            == 200.0
+        )
+
+    with subtests.test("Min LCOH electrolyzer capacity"):
+        assert (
+            pytest.approx(
+                cases[min_lcoh_case_num].get_val("electrolyzer.electrolyzer_size_mw", units="MW"),
+                rel=1e-6,
+            )
+            == 100.0
+        )
+
+    # remove files created
+    new_driver_fpath.unlink()
+    new_toplevel_fpath.unlink()
+    new_csv_filename.unlink()