+++ CHANGELOG.md
@@ -4,6 +4,22 @@
 - Added `tools/run_cases.py` with tools to run different `tech_config` cases from a spreadsheet, with new docs page to describe: docs/user_guide/how_to_run_several_cases_in_sequence.md
 - Added capability for user-defined finance models in the H2Integrate framework
 - Added an optimized offshore methanol production case to examples/03_methanol/co2_hydrogenation_doc
+- Updated setting up recorder in `PoseOptimization`
+- Added resource models to make solar resource API calls to the NREL Developer GOES dataset
+- Added framework to run heuristic load following dispatch for storage technologies
+- Added PySAM battery model as a storage technology performance model
+- Added `create_om_reports` option to driver config to enable/disable OpenMDAO reports (N2 diagrams, etc.)
+- Added design of experiment functionality
+- Added "csvgen" as generator type for design of experiments
+- Added PySAM Windpower performance model to simulate wind.
+- Added `simple_grid_layout.py` for wind plant layout modeling, can model square or rectangular layouts.
+- Added ability to visualize the wind plant layout for PySAM Windpower model using `post_process(show_plots=True)`.
+- Added Wind Annual Technology Baseline cost model `atb_wind_cost.py`.
+- Updated inputs for the `ATBBatteryCostModel` and `DemandOpenLoopController` so storage capacity and charge rate can be design variables
+- Created `ProFastBase`, a base class for the `ProFastLCO` and `ProFastNPV` models
+- Added `ProFastNPV`, a finance model using ProFAST to calculate NPV of the commodity
+- Moved `compute()` from `ProFastBase` to `ProFastLCO`.
+- Added `NumpyFinancialNPV`, a finance model that uses NumPy Financial npv to calculate the npv from the cash flows
 
 ## 0.4.0 [October 1, 2025]